requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
tk>=0.1.0
ttkbootstrap>=1.10.0